pyproject.toml
[project]
name = "movalytics-pipeline"
version = "0.1.0"
description = "3D human pose estimation from monocular video: YOLOX detection → SAM 3D Body → clinical joint angles + OpenSim IK"
readme = "README.md"
license = {text = "CC-BY-NC-SA-4.0"}
requires-python = "==3.12.*"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pillow>=10.2.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "torch>=2.9.1",
    "rtmlib>=0.0.10",
    "tqdm>=4.60.0",
    "onnxruntime>=1.17.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]

[tool.uv]
override-dependencies = [
    "pyqt5 ; sys_platform != 'win32'",
    "pyqt5-qt5 ; sys_platform != 'win32'",
    "pyqt5-sip ; sys_platform != 'win32'",
]
